refactor(#5031): W2 extract skills_scan from discord/mod.rs (move-only)
